requirements.txt
Flask
python-dotenv
altcha